requirements.txt
phidata
python-dotenv~=1.0.1
yfinance
packaging~=23.2
duckduckgo-search
fastapi~=0.115.8
uvicorn~=0.34.0
groq
openai~=1.60.2
sqlalchemy~=2.0.37
pgvector
psycopg[binary]~=3.2.4
pypdf~=4.3.1
streamlit~=1.42.0
phidata
google-generativeai
duckduckgo-search
playwright~=1.50.0
lxml~=5.3.0
bs4~=0.0.2
beautifulsoup4~=4.13.3
soupsieve~=2.6
html5lib~=1.1
h11~=0.14.0
cryptography~=44.0.1
cffi~=1.17.1
defusedxml~=0.7.1
numpy~=1.26.4
pillow~=10.4.0
ipython~=7.34.0
pip~=21.3.1
attrs~=25.1.0
argparse~=1.4.0
distro~=1.9.0
wheel~=0.37.1
tornado~=6.4.2
jinja2~=3.1.5
rsa~=4.9
pyasn1~=0.6.1
setuptools~=75.8.0
parso~=0.8.4
jedi~=0.19.2
colorama~=0.4.6
msal~=1.31.1
nltk~=3.9.1
matplotlib~=3.10.0
regex~=2024.11.6
scipy~=1.15.1
scikit-learn~=1.6.1
click~=8.1.8
tqdm~=4.67.1
requests~=2.32.3
pyparsing~=3.2.1
pytz~=2025.1
rich~=13.9.4
pygments~=2.19.1
trio~=0.29.0
outcome~=1.3.0.post0
sniffio~=1.3.1
sortedcontainers~=2.4.0
idna~=3.10
PyYAML~=6.0.2
multidict~=6.1.0
propcache~=0.2.1
aiohttp~=3.11.12
urllib3~=2.3.0
babel~=2.17.0
ffmpy~=0.5.0
httpcore~=1.0.7
httpx~=0.28.1
zstandard~=0.23.0
pywin32~=308
jiter~=0.8.2
mdurl~=0.1.2
Cython~=3.0.11
typer~=0.15.1
shellingham~=1.5.4
fsspec~=2024.12.0
yarl~=1.18.3
cachetools~=5.5.1
certifi~=2025.1.31
google-api-python-client~=2.160.0
pydantic~=2.10.6
pandas~=2.2.3
gradio~=4.39.0
tomlkit~=0.12.0
anyio~=4.8.0
orjson~=3.10.15
MarkupSafe~=2.1.5
python-multipart~=0.0.20
starlette~=0.45.3
filelock~=3.17.0
aiofiles~=23.2.1
pydub~=0.25.1
joblib~=1.4.2
websockets~=11.0.3
pyarrow~=19.0.0
python-dateutil~=2.9.0.post0
aiosignal~=1.3.2
frozenlist~=1.5.0
aiohappyeyeballs~=2.4.4
tld~=0.13
traitlets~=5.14.3
decorator~=5.1.1
backcall~=0.2.0
pickleshare~=0.7.5
wcwidth~=0.2.13
lavague~=1.1.19
msgpack~=1.1.0
selenium~=4.28.1
seaborn~=0.13.2
jsonschema~=4.23.0
langchain~=0.1.20
trafilatura~=1.12.2
tiktoken~=0.8.0
cycler~=0.12.1
tzlocal~=5.2
wsproto~=1.2.0
six~=1.17.0
greenlet~=3.1.1
dateparser~=1.2.1
networkx~=3.4.2
sympy~=1.13.1
Deprecated~=1.2.18
chardet~=5.2.0
websocket-client~=1.8.0
tenacity~=8.3.0
contourpy~=1.3.1
dirtyjson~=1.0.8
fonttools~=4.56.0
reportlab~=4.3.1
langsmith~=0.1.147
pycparser~=2.22
wrapt~=1.17.2
referencing~=0.36.2
jsonpointer~=3.0.0
kiwisolver~=1.4.8
zipp~=3.21.0
botocore~=1.36.14
torch~=2.6.0
marshmallow~=3.26.1
courlan~=1.3.2
htmldate~=1.9.3
jusText~=3.0.1
jsonpatch~=1.33
pyperclip~=1.9.0
portalocker~=2.10.1
boto3~=1.36.14
opencv-python~=4.11.0.86
phi~=0.6.7
pypdf2~=3.0.1
fillpdf~=0.7.3